requirements.txt
python-dotenv==1.0.1
uvicorn==0.32.0
fastapi==0.115.2
sqlmodel==0.0.22
Jinja2==3.1.4
python-docx==1.1.2
python-multipart==0.0.16
openai==1.52.2
